Refactor conditionfunc to use ParseCommitInfo
